Relax httparty dependency restriction
Given that https://github.com/jnunemaker/httparty/issues/733 is
resolved, this removes the restriction to stay on the 0.18 minor version
of httparty.

Notably, the intent here is to allow for httparty to be updated in
light of this issue: https://github.com/advisories/GHSA-5pq7-52mg-hr42.

httparty 0.21.0 has been released with a fix for this issue.